transferred eveyrthing from server.js to index.js
